Fix for deadlock after stopping and restarting IOCs.
Problem: The methods
org.csstudio.alarm.diirt.datasource.BeastDataSource.BeastAlarmClientModelListener.newAlarmState
and org.csstudio.alarm.diirt.datasource.MultiplexedChannelHandler.removeWrite
and their subcalls compete for a lock on a
BeastDataSource.channelConsumers object and a BeastChannelHandler object
at the same time, getting a deadlock.

Solution: Narrow the critical section in method newAlarmState when reconnecting the channels and avoiding to lock on channelConsumers map. As far the channelConsumers map is thread-safe, we can get its keys atomically and reconnect each one later in a loop (it is a snapshot). If in the middle of the loop, one of the channels is dropped from the collection, the object will still be alive  and accessible, so it will reconnect, although eventually it will be no longer available and reclaimed by GC (anyway the method disconnect of the BeastChannelHandler does nothing more than to remove the channel from the collection).