Fix for deadlock after stopping and restarting IOCs.
Problem: The methods
org.csstudio.alarm.diirt.datasource.BeastDataSource.remove,
org.csstudio.alarm.diirt.datasource.MultiplexedChannelHandler.processMessage,
and their subcalls compete for a lock on a
BeastDataSource.channelConsumers object and a BeastChannelHandler object
at the same time, getting a deadlock.

Solution: Narrow the critical section in method remove when removing the
consumer avoiding to lock on channelConsumers map. As far the
channelConsumers map is thread-safe, we can get its channel atomically
and remove safely a consumer from it (the channel is a list) using a
synchronized section. If two different threads get the same channel and
try to remove the same consumer (sequentially because is protected by
the synchronized section), the second call to channel.remove will return
a false which was already ignored anyway. If the channel itself is
removed from channelConsumers map by another thread, the current thread
will remove a consumer from a list which is not used, but this is
harmless.

Note: It has also been improved the call to get the channel. Is not necessary to ask first if it exists in the map, instead the call to method get will return a null if it does not exist.